增加了cell高度自适应＋详细项目页面
增加了cell高度自适应＋详细项目页面

+++ BugBox/BugBox/Classes/MainControl/addressBook/AddressBookController.m
@@ -79,7 +79,7 @@
     static NSInteger times = 0;
     [[LeanMessageManager manager] openSessionWithClientID:clientId completion:^(BOOL succeeded, NSError *error) {
         times++;
-        NSLog(@"times = %ld", times);
+        NSLog(@"times = %ld", (long)times);
         if(!error){
             ConversationType type;
             if(clientIDs.count>1){
